无语的return

+++ models/Result.go
@@ -1,7 +1,7 @@
 package models
 
 type Error struct {
-	Status byte
+	Status int
 	Msg    string
 }
 
@@ -9,11 +9,11 @@ func ReurnError(status int,msg string) *Error {
 	if msg == "" {
 		msg = "error"
 	}
-	return &Error{Status: byte(status), Msg: msg}
+	return &Error{Status: status, Msg: msg}
 }
 
 type Success struct {
-	Status byte
+	Status int
 	Msg    string
 }
 
@@ -21,5 +21,5 @@ func ReurnSuccess(msg string) *Error {
 	if msg == "" {
 		msg = "success"
 	}
-	return &Error{Status: byte(0), Msg: msg}
+	return &Error{Status: 0, Msg: msg}
 }
